update the folder
