+++ src/App.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import Layout from './components/design';
-/* eslint-disable */
 function App() {
   return (
     <div className="App">
